+++ src/main/java/com/myce/member/controller/MemberMyPageController.java
@@ -1,0 +1,94 @@
+package com.myce.member.controller;
+
+import com.myce.auth.dto.CustomUserDetails;
+import com.myce.member.dto.*;
+import com.myce.member.service.MemberMyPageService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/api/members")
+@RequiredArgsConstructor
+public class MemberMyPageController {
+    
+    private final MemberMyPageService memberMyPageService;
+    
+    @GetMapping("/my-page/reserved-expos")
+    public ResponseEntity<List<ReservedExpoResponse>> getReservedExpos(
+            @AuthenticationPrincipal CustomUserDetails customUserDetails) {
+        
+        Long memberId = customUserDetails.getMemberId();
+        List<ReservedExpoResponse> reservedExpos = memberMyPageService.getReservedExpos(memberId);
+        
+        return ResponseEntity.ok(reservedExpos);
+    }
+
+    @GetMapping("/my-page/favorite_expos")
+    public ResponseEntity<List<FavoriteExpoResponse>> getFavoriteExpos(
+            @AuthenticationPrincipal CustomUserDetails customUserDetails) {
+
+        Long memberId = customUserDetails.getMemberId();
+        List<FavoriteExpoResponse> favoriteExpos = memberMyPageService.getFavoriteExpos(memberId);
+
+        return ResponseEntity.ok(favoriteExpos);
+    }
+    
+    @GetMapping("/my-page/info")
+    public ResponseEntity<MemberInfoResponse> getMemberInfo(
+            @AuthenticationPrincipal CustomUserDetails customUserDetails) {
+        
+        Long memberId = customUserDetails.getMemberId();
+        MemberInfoResponse memberInfo = memberMyPageService.getMemberInfo(memberId);
+        
+        return ResponseEntity.ok(memberInfo);
+    }
+    
+    @PutMapping("/my-page/info")
+    public ResponseEntity<Void> updateMemberInfo(
+            @AuthenticationPrincipal CustomUserDetails customUserDetails,
+            @RequestBody MemberInfoUpdateRequest request) {
+        
+        Long memberId = customUserDetails.getMemberId();
+        memberMyPageService.updateMemberInfo(memberId, request);
+        
+        return ResponseEntity.ok().build();
+    }
+
+    @GetMapping("/my-page/payment-history")
+    public ResponseEntity<Page<PaymentHistoryResponse>> getPaymentHistory(
+            @AuthenticationPrincipal CustomUserDetails customUserDetails,
+            Pageable pageable) {
+        
+        Long memberId = customUserDetails.getMemberId();
+        Page<PaymentHistoryResponse> paymentHistory = memberMyPageService.getPaymentHistory(memberId, pageable);
+        
+        return ResponseEntity.ok(paymentHistory);
+    }
+    
+    @GetMapping("/my-page/settings")
+    public ResponseEntity<MemberSettingResponse> getMemberSetting(
+            @AuthenticationPrincipal CustomUserDetails customUserDetails) {
+        
+        Long memberId = customUserDetails.getMemberId();
+        MemberSettingResponse memberSetting = memberMyPageService.getMemberSetting(memberId);
+        
+        return ResponseEntity.ok(memberSetting);
+    }
+    
+    @PutMapping("/my-page/settings")
+    public ResponseEntity<Void> updateMemberSetting(
+            @AuthenticationPrincipal CustomUserDetails customUserDetails,
+            @RequestBody MemberSettingUpdateRequest request) {
+        
+        Long memberId = customUserDetails.getMemberId();
+        memberMyPageService.updateMemberSetting(memberId, request);
+        
+        return ResponseEntity.ok().build();
+    }
+}
